+++ user-management-service/usermanagement-api/gradle.properties
@@ -1,0 +1,2 @@
+apiVersion=0.1-SNAPSHOT
+releaseTrainVersion=Greenwich.SR2
