+++ user-management-service/usermanagement-api/src/main/java/com/microservices/usermanagement/api/dto/CardInfo.java
@@ -1,0 +1,6 @@
+package com.microservices.usermanagement.api.dto;
+
+public enum CardInfo {
+    VALID,
+    INVALID
+}
