+++ user-management-service/usermanagement-api/src/main/java/com/microservices/usermanagement/api/service/UserServiceApi.java
@@ -1,0 +1,17 @@
+package com.microservices.usermanagement.api.service;
+
+import com.microservices.usermanagement.api.dto.UserDto;
+
+/**
+ * Interface for accessing user service
+ */
+public interface UserServiceApi {
+
+    /**
+     * Returns saved user by provided login
+     *
+     * @throws IllegalArgumentException if specified used is not found
+     * @throws NullPointerException     if provided login is null
+     */
+    UserDto getUserByLogin(String login);
+}
