+++ user-management-service/usermanagement-api/src/main/resources/application.properties
@@ -1,0 +1,2 @@
+#The name for user service that is used by Eureka and Ribbon (in Feign)
+user.management.service.name=user-management-service
