+++ user-management-service/usermanagement-app/build.gradle
@@ -21,10 +21,12 @@ bootJar {
 sourceCompatibility = 1.8
 
 repositories {
+    mavenLocal()
     mavenCentral()
 }
 
 dependencies {
+    compile "com.microservices.usermanagement.api:usermanagementapi:$userManagementApiVersion"
     compile 'org.springframework.boot:spring-boot-starter-web'
     compile 'org.springframework.boot:spring-boot-starter-data-jpa'
     compile 'org.flywaydb:flyway-core:5.2.4'
