+++ user-management-service/usermanagement-app/gradle.properties
@@ -1,1 +1,2 @@
 applicationVersion = 0.1-SNAPSHOT
+userManagementApiVersion=0.1-SNAPSHOT
