+++ user-management-service/usermanagement-app/src/main/java/com/microservices/usermanagement/controller/UserController.java
@@ -1,11 +1,17 @@
 package com.microservices.usermanagement.controller;
 
+import com.microservices.usermanagement.api.dto.UserDto;
 import com.microservices.usermanagement.dto.CreateUserDto;
-import com.microservices.usermanagement.dto.UserDto;
 import com.microservices.usermanagement.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseStatus;
+import org.springframework.web.bind.annotation.RestController;
 
 import javax.validation.Valid;
