+++ user-management-service/usermanagement-app/src/main/java/com/microservices/usermanagement/dto/CreateUserDto.java
@@ -1,6 +1,6 @@
 package com.microservices.usermanagement.dto;
 
-import com.microservices.usermanagement.entity.CardInfo;
+import com.microservices.usermanagement.api.dto.CardInfo;
 import lombok.Getter;
 import lombok.Setter;
 import lombok.ToString;
