--- user-management-service/usermanagement-app/src/main/java/com/microservices/usermanagement/entity/CardInfo.java
@@ -1,6 +1,0 @@
-package com.microservices.usermanagement.entity;
-
-public enum CardInfo {
-    VALID,
-    INVALID
-}
