+++ user-management-service/usermanagement-app/src/main/java/com/microservices/usermanagement/entity/UserEntity.java
@@ -1,11 +1,19 @@
 package com.microservices.usermanagement.entity;
 
+import com.microservices.usermanagement.api.dto.CardInfo;
+import com.microservices.usermanagement.api.dto.UserDto;
 import com.microservices.usermanagement.dto.CreateUserDto;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-import javax.persistence.*;
+import javax.persistence.Entity;
+import javax.persistence.EnumType;
+import javax.persistence.Enumerated;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.Table;
 import java.util.Objects;
 
 @Getter
@@ -27,5 +35,14 @@ public class UserEntity {
         Objects.requireNonNull(createUserDto);
         this.login = createUserDto.getLogin();
         this.cardInfo = createUserDto.getCardInfo();
+    }
+
+
+    public UserDto toUserDto() {
+        UserDto userDto = new UserDto();
+        userDto.setId(id);
+        userDto.setLogin(login);
+        userDto.setCardInfo(cardInfo);
+        return userDto;
     }
 }
