+++ user-management-service/usermanagement-app/src/main/java/com/microservices/usermanagement/service/UserService.java
@@ -1,7 +1,7 @@
 package com.microservices.usermanagement.service;
 
+import com.microservices.usermanagement.api.dto.UserDto;
 import com.microservices.usermanagement.dto.CreateUserDto;
-import com.microservices.usermanagement.dto.UserDto;
 
 public interface UserService {
     UserDto getUserByLogin(String login);
