+++ user-management-service/usermanagement-app/src/main/resources/application.properties
@@ -1,6 +1,7 @@
 # The port that will be used for REST access
 server.port=8094
-
+# Application name is used by Eureka Service Discovery, identifying traces in Zipkin
+spring.application.name=user-management-service
 # Database and Hibernate properties
 database.schema.name=user_management
 spring.datasource.url=jdbc:postgresql://localhost:5432/laundry_db
